+++ auth-db/queries/email-auth/index.ts
@@ -8,3 +8,4 @@ export * from "./update-password.ts";
 export * from "./clear-forgot-password-token.ts";
 export * from "./verify-email.ts";
 export * from "./get-by-verification-token.ts";
+export * from "./update-verification-token.ts";
